% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simon_ae2_attend.R
\name{ae2_attendances}
\alias{ae2_attendances}
\title{ae2_attendances}
\usage{
ae2_attendances(AE2, target_diagnosis)
}
\arguments{
\item{AE2}{data.frame including AE2 records}

\item{target_diagnosis}{integer specifying the target diagnosis number}
}
\value{
feature_df data.frame with id and num_ae2_attendances columns
}
\description{
Computes the features including the count of AE2 records per
id, number of drug and alcohol attendances and number of pysch attendances
}
