% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{check_for_nested}
\alias{check_for_nested}
\title{Check if a filter is nested or single}
\usage{
check_for_nested(filter)
}
\arguments{
\item{filter}{A filter as defined in the origin json file}
}
\description{
Check if a filter is nested or single
}
