% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_ae2.R
\name{clean_ae2}
\alias{clean_ae2}
\title{Cleans the AE2 table. Adds the attendance_category column and populates it
with 1 if it doesn't exist.}
\usage{
clean_ae2(ae2, patient_id_col = "unique_study_id")
}
\arguments{
\item{ae2}{data.frame of ae2 records to be cleaned. The following columns must exist:
admission_date, diagnosis_1, diagnosis_2, diagnosis_3, as well as the column
referred to in patient_id_col}

\item{patient_id_col}{name of column in ae2 containing patient IDs. Defaults
to "unique_study_id"}
}
\value{
data.frame of cleaned ae2 data
}
\description{
Cleans the AE2 table. Adds the attendance_category column and populates it
with 1 if it doesn't exist.
}
