% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ae2.R
\name{clean_ae2_jl}
\alias{clean_ae2_jl}
\title{James's version of AE2 cleaning, tidied up to avoid using files.
Mostly kept here for posterity.}
\usage{
clean_ae2_jl(table_ae2, patient_id_col = "unique_study_id")
}
\arguments{
\item{patient_id_col}{name of column in ae2 containing patient IDs. Defaults
to "unique_study_id"}

\item{table_AE2}{data.frame of AE2 records to be cleaned. The following
column names must be present: unique_study_id, admission_date,
admission_time, transfer_discharge_date, transfer_discharge_time.}
}
\value{
data.frame of cleaned AE2 data
}
\description{
James's version of AE2 cleaning, tidied up to avoid using files.
Mostly kept here for posterity.
}
