% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_smr00.R
\name{clean_rawData}
\alias{clean_rawData}
\title{Clean raw data and generate tidy versions of data tables. Will only update episodes table if force_redo = TRUE.}
\usage{
clean_rawData(force_redo = TRUE, dir_rawData = ".", dir_cleanData = ".")
}
\arguments{
\item{force_redo}{Redo if clean data are not already there}

\item{dir_rawData}{directory containing raw data tables (from EHRs)}

\item{dir_cleanData}{directory to which clean data tables will be written}
}
\description{
Clean raw data and generate tidy versions of data tables. Will only update episodes table if force_redo = TRUE.
}
