% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_ae2.R
\name{count_ae2_attendance_features}
\alias{count_ae2_attendance_features}
\title{Computes features from the AE2 table, including the count of AE2 records per
id, number of drug and alcohol attendances, and number of psych attendances.}
\usage{
count_ae2_attendance_features(ae2)
}
\arguments{
\item{ae2}{Cleaned AE2 data.frame. Must include the following columns:
id, attendance_category, diagnosis_1, diagnosis_2, diagnosis_3}
}
\value{
data.frame with th efollowing columns:
id, num_ae2_attendances, num_alcohol_drug_attendances, num_psych_attendances
}
\description{
Computes features from the AE2 table, including the count of AE2 records per
id, number of drug and alcohol attendances, and number of psych attendances.
}
