% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_ae2.R
\name{count_matching_diagnoses}
\alias{count_matching_diagnoses}
\title{Computes the number of AE2 attendances with a given diagnosis.}
\usage{
count_matching_diagnoses(ae2, target_diagnoses, output_column_name)
}
\arguments{
\item{ae2}{Cleaned AE2 data.frame.}

\item{target_diagnoses}{Vector of diagnosis codes to count.}
}
\description{
Computes the number of AE2 attendances with a given diagnosis.
}
