% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise.R
\name{featurise}
\alias{featurise}
\title{Extract a feature table using a feature JSON file.}
\usage{
featurise(all_tables, feature_json_file, context = NULL)
}
\arguments{
\item{all_tables}{List of all input tables (passed in from read_all_tables).}

\item{feature_json_file}{Path to the feature JSON file.}

\item{context}{A character vector to be used in logging or error messages.
Defaults to NULL.}
}
\value{
A list with the following elements:
\itemize{
\item feature_table: A data frame with one row per patient ID and one column
containing the sum of matching rows in the source table.
The column names are 'id' for the ID (this is standardised
across all feature tables), and the value of
output_column_name.
\item missing_value: The value to use for patients who have no matching rows in
the source table. This value is passed downstream to the
function which joins all the feature tables together.
}
}
\description{
Extract a feature table using a feature JSON file.
}
