% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_combine_linear.R
\name{featurise_combine_linear}
\alias{featurise_combine_linear}
\title{Computes the linear combination of two or more features.}
\usage{
featurise_combine_linear(all_tables, spec, context = NULL)
}
\arguments{
\item{all_tables}{List of all input tables (passed in from read_all_tables).}

\item{spec}{A list containing the following elements:
\itemize{
\item output_feature_name: Name of the output column.
\item grouping_columns:    Name of the column(s) to group by.
\item feature_list:        List of feature specs to combine.
}}

\item{context}{A character vector to be used in logging or error messages.
Defaults to NULL.}
}
\value{
A list with the following elements:
\itemize{
\item feature_table: A data frame with one row per patient ID and one column
containing the count of matching rows in the source table.
The column names are 'id' for the ID (this is standardised
across all feature tables), and the value of
output_column_name.
\item missing_value: The value to use for patients who have no matching rows in
the source table. This value is passed downstream to the
function which joins all the feature tables together.
}
}
\description{
Computes the linear combination of two or more features.
}
