% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_count.R
\name{featurise_count}
\alias{featurise_count}
\title{Computes the number of rows per patient ID which contain specific values in
specific columns.}
\usage{
featurise_count(
  all_tables,
  source_table_file,
  filter_obj,
  id_column_name = "id",
  missing_value = 0,
  output_column_name
)
}
\arguments{
\item{all_tables}{List of all input tables (passed in from read_all_tables).}

\item{source_table_file}{Filename of the source table to read from.}

\item{filter_obj}{A filter object to apply to the source table.}

\item{id_column_name}{Name of the patient ID column in the source table.}

\item{missing_value}{The value to use for patients who have no matching rows
in the source table.}

\item{output_column_name}{Name of the output column.}
}
\value{
A list with the following elements:
\itemize{
\item feature_table: A data frame with one row per patient ID and one column
containing the count of matching rows in the source table.
The column names are 'id' for the ID (this is standardised
across all feature tables), and the value of
output_column_name.
\item missing_value: The value to use for patients who have no matching rows in
the source table. This value is passed downstream to the
function which joins all the feature tables together.
}
}
\description{
Computes the number of rows per patient ID which contain specific values in
specific columns.
}
