% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_count_per_id.R
\name{featurise_count_per_id}
\alias{featurise_count_per_id}
\title{Computes the number of rows per patient ID which contain a specific
diagnosis code.}
\usage{
featurise_count_per_id(ae2, target_diagnoses, output_column_name)
}
\arguments{
\item{ae2}{Cleaned AE2 data.frame.}

\item{target_diagnoses}{Vector of diagnosis codes to count.

TODO: Generalise over the following:
\itemize{
\item 'diagnosis_x' column names
\item 'id' column name
\item source table name
}}
}
\description{
Computes the number of rows per patient ID which contain a specific
diagnosis code.
}
