% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_time_since.R
\name{featurise_time_since}
\alias{featurise_time_since}
\title{First passes a table through a filter, then counts the time between a given
date and the first OR last date in the filtered table.}
\usage{
featurise_time_since(
  all_tables,
  source_table_file,
  filter_obj,
  date_column_name,
  cutoff_date,
  from_first,
  time_units,
  output_column_name,
  id_column_name = "id",
  missing_value = 0
)
}
\arguments{
\item{all_tables}{List of all input tables (passed in from read_all_tables).}

\item{source_table_file}{Filename of the source table to read from.}

\item{filter_obj}{A filter object to apply to the source table.}

\item{date_column_name}{Name of the date column in the source table to
calculate the time since.}

\item{cutoff_date}{The date to calculate the time since from.}

\item{from_first}{If TRUE, calculate the time since the first date in the
filtered table. If FALSE, calculate the time since the last (most recent)
date in the filtered table.}

\item{time_units}{Either "days" or "years". The number of years is defined as
the number of full 365.25 day periods between the two dates (and is rounded
down to the nearest whole number).}

\item{output_column_name}{Name of the output column.}

\item{id_column_name}{Name of the patient ID column in the source table.
Defaults to 'id'.}

\item{missing_value}{The value to use for patients who have no matching rows
in the source table. Defaults to 0.}
}
\value{
A list with the following elements:
\itemize{
\item feature_table: A data frame with one row per patient ID and one column
containing the count of matching rows in the source table.
The column names are 'id' for the ID (this is standardised
across all feature tables), and the value of
output_column_name.
\item missing_value: The value to use for patients who have no matching rows in
the source table. This value is passed downstream to the
function which joins all the feature tables together.
}
}
\description{
First passes a table through a filter, then counts the time between a given
date and the first OR last date in the filtered table.
}
