% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_unique.R
\name{featurise_unique}
\alias{featurise_unique}
\title{Groups by ID, filters for rows that obey certain conditions, and computes
the number of unique values in a given column.}
\usage{
featurise_unique(
  all_tables,
  source_table_file,
  filter_obj,
  aggregate_column_name,
  output_column_name,
  id_column_name = "id",
  missing_value = 0
)
}
\arguments{
\item{all_tables}{List of all input tables (passed in from read_all_tables).}

\item{source_table_file}{Filename of the source table to read from.}

\item{filter_obj}{A filter object to apply to the source table.}

\item{aggregate_column_name}{Name of the column over which to aggregate.}

\item{output_column_name}{Name of the output column.}

\item{id_column_name}{Name of the patient ID column in the source table.
Defaults to 'id'.}

\item{missing_value}{The value to use for patients who have no matching rows
in the source table. Defaults to 0.}
}
\value{
A list with the following elements:
\itemize{
\item feature_table: A data frame with one row per patient ID and one column
containing the sum of matching rows in the source table.
The column names are 'id' for the ID (this is standardised
across all feature tables), and the value of
output_column_name.
\item missing_value: The value to use for patients who have no matching rows in
the source table. This value is passed downstream to the
function which joins all the feature tables together.
}
}
\description{
Groups by ID, filters for rows that obey certain conditions, and computes
the number of unique values in a given column.
}
