% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_all.R
\name{filter_all}
\alias{filter_all}
\title{Pass table through any type of filter.}
\usage{
filter_all(table, filter_obj)
}
\arguments{
\item{table}{A table to filter.}

\item{filter_obj}{A filter to apply to the table. This is a list
which can have 'type' = any of 'basic', 'or', 'and', or 'not'.}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Pass table through any type of filter.
}
