% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and.R
\name{filter_and}
\alias{filter_and}
\title{Filter table rows based on a logical conjunction of multiple filters (i.e.
filter1 AND filter2 AND ...)}
\usage{
filter_and(table, filter_obj, context = NULL)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'and'
- subfilters: a list of filter objects}

\item{context}{A string to be used in logging or error messages. Defaults to
NULL.}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Filter table rows based on a logical conjunction of multiple filters (i.e.
filter1 AND filter2 AND ...)
}
