% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_basic_date.R
\name{filter_basic_date}
\alias{filter_basic_date}
\title{Generic function for a basic filter, parametrised over the type of
comparison operator used to select rows. The values to be taken are
dates instead of numbers, though.}
\usage{
filter_basic_date(table, filter_obj)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'in', 'less_than', 'less_than_equal',
'greater_than', or 'greater_than_equal'
- column: the name of the column to filter on
- value: a vector of values to filter on (for type 'in'), or
a single value (for all other types)}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Generic function for a basic filter, parametrised over the type of
comparison operator used to select rows. The values to be taken are
dates instead of numbers, though.
}
