% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_in.R
\name{filter_in}
\alias{filter_in}
\title{Filter table rows based on a single condition, i.e. a column having values
in a given set}
\usage{
filter_in(table, filter_obj)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'basic'
- column: the name of the column to filter on
- values: a vector of values to filter on}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Filter table rows based on a single condition, i.e. a column having values
in a given set
}
