% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_less_than.R
\name{filter_less_than}
\alias{filter_less_than}
\title{Filter table rows based on a comparison to a given value. All rows with a value
less than the given value are included in the result. The comparison is
exclusive i.e. rows with exactly this value are not included in the result.}
\usage{
filter_less_than(table, filter_obj)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'less_than'
- column: the name of the column to filter on
- value: the threshold to filter on}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Filter table rows based on a comparison to a given value. All rows with a value
less than the given value are included in the result. The comparison is
exclusive i.e. rows with exactly this value are not included in the result.
}
