% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and.R, R/filter_or.R
\name{filter_or}
\alias{filter_or}
\title{Filter table rows based on a logical conjunction of multiple filters (i.e.
filter1 AND filter2 AND ...)}
\usage{
filter_or(table, filter_obj)

filter_or(table, filter_obj)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'or'
- subfilters: a list of filter objects}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows

A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Filter table rows based on a logical conjunction of multiple filters (i.e.
filter1 AND filter2 AND ...)

Filter table rows based on a logical disjunction of multiple filters (i.e.
filter1 OR filter2 OR ...)
}
