% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{join_feature_df}
\alias{join_feature_df}
\title{Used when we are adding sets of features one at a time to a single
data.frame}
\usage{
join_feature_df(current_df, new_df, join_by = "id")
}
\arguments{
\item{current_df}{data.frame with the current features (must include id
column). This can be NULL to start with.}

\item{new_df}{data.frame with the new features (must include id column). If
current_df is NULL current_df is assigned to new_df, otherwise a full join
is performed.}

\item{join_by}{column name to join by}
}
\value{
data.frame formed by joining
}
\description{
Used when we are adding sets of features one at a time to a single
data.frame
}
