% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_feature_tables.R
\name{join_feature_tables}
\alias{join_feature_tables}
\title{Join feature tables together.}
\usage{
join_feature_tables(calculated_features)
}
\arguments{
\item{calculated_features}{A list of calculated features, each of which is
produced by the featurise_... functions}
}
\value{
A data frame with the feature tables joined together
}
\description{
Join feature tables together.
}
