% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{json_to_feature}
\alias{json_to_feature}
\title{Take the json input file and produce a formatted R object which can direct
the filters}
\usage{
json_to_feature(filename)
}
\arguments{
\item{filename}{The relative filepath to the json file}
}
\value{
A feature object
}
\description{
Take the json input file and produce a formatted R object which can direct
the filters
}
