% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw_ae2.R
\name{make_raw_ae2}
\alias{make_raw_ae2}
\title{Generate a raw AE2 table.}
\usage{
make_raw_ae2(n = 6, nid = 15, seed = 468)
}
\arguments{
\item{n}{Number of rows to generate.}

\item{nid}{Number of unique study IDs to generate.}

\item{seed}{Seed for random number generation.}
}
\description{
Generate a raw AE2 table.
}
