% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniSPARRA01-package.R
\docType{package}
\name{miniSPARRA01-package}
\alias{miniSPARRA01}
\alias{miniSPARRA01-package}
\title{miniSPARRA01: What the Package Does (One Line, Title Case)}
\description{
Minisparra.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
