% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{minisparra_example}
\alias{minisparra_example}
\title{Return path to the example datasets provided with the package}
\usage{
minisparra_example(file = NULL)
}
\arguments{
\item{file}{Name of the file to be used from inst/extdata. If NULL all files will be listed}
}
\description{
Return path to the example datasets provided with the package
}
\examples{
minisparra_example() Returns a list of all the random patient data
minisparra_example("random_ae_data.csv") Returns path to the random A&E type file
}
