% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_header_info}
\alias{parse_header_info}
\title{Parse the header information from the json file to our targed feature object}
\usage{
parse_header_info(feature_object, json_data)
}
\arguments{
\item{feature_object}{The R object which will define the filtering}

\item{json_data}{The parsed json data}
}
\value{
A feature object
}
\description{
Parse the header information from the json file to our targed feature object
}
