% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_single_feature}
\alias{parse_single_feature}
\title{Parse the header information from the json file to our targed feature object}
\usage{
parse_single_feature(json_data)
}
\arguments{
\item{json_data}{The parsed json data}
}
\value{
A feature object
}
\description{
Parse the header information from the json file to our targed feature object
}
