% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_single_filter}
\alias{parse_single_filter}
\title{Parse a single (un-nested) filter and return a list with data column name
filter type, and the limiting values}
\usage{
parse_single_filter(filter)
}
\arguments{
\item{filter}{A filter as defined in the origin json file}
}
\description{
Parse a single (un-nested) filter and return a list with data column name
filter type, and the limiting values
}
