% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_tables.R
\name{read_all_tables}
\alias{read_all_tables}
\title{Read in all tables from a list of filenames.}
\usage{
read_all_tables(filenames)
}
\arguments{
\item{filenames}{A vector of filenames.}
}
\value{
A list of data frames.
}
\description{
Read in all tables from a list of filenames.
}
