% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_context.R
\name{trace_context}
\alias{trace_context}
\title{Log a TRACE message showing the current execution context}
\usage{
trace_context(context)
}
\arguments{
\item{context}{A character vector}
}
\description{
Log a TRACE message showing the current execution context
}
