% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Perform the entire feature transformation process.}
\usage{
transform(all_table_filenames, all_feature_json_filenames)
}
\arguments{
\item{all_table_filenames}{A vector of file paths to the source tables.}

\item{all_feature_json_filenames}{A vector of file paths to the feature JSON
specifications.}
}
\value{
A data frame with the feature tables joined together
}
\description{
Perform the entire feature transformation process.
}
