% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_all.R
\name{filter_all}
\alias{filter_all}
\title{Pass table through any type of filter.}
\usage{
filter_all(table, filter_obj, context = NULL)
}
\arguments{
\item{table}{A table to filter.}

\item{filter_obj}{A filter to apply to the table.}

\item{context}{A string to be used in logging or error messages. Defaults to
NULL.}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Pass table through any type of filter.
}
