% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_basic.R
\name{filter_basic}
\alias{filter_basic}
\title{Generic function for a basic filter, parametrised over the type of
comparison operator used to select rows.}
\usage{
filter_basic(table, filter_obj, context = NULL)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'in', 'lt', 'lt_eq', 'gt', or 'gt_eq' (case-
insensitive)
- column: the name of the column to filter on
- value: a vector of values to filter on (for type 'in'), or
a single value (for all other types)}

\item{context}{A string to be used in logging or error messages. Defaults to
NULL.}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Generic function for a basic filter, parametrised over the type of
comparison operator used to select rows.
}
