% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_not.R
\name{filter_not}
\alias{filter_not}
\title{Filter table rows based on a logical negation of a filter (i.e. NOT filter1)}
\usage{
filter_not(table, filter_obj, context = NULL)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'not' (case-insensitive)
- subfilter: a single filter object}

\item{context}{A string to be used in logging or error messages. Defaults to
NULL.}
}
\value{
A list with the following elements:
- passed: data frame with the rows that satisfies the NOT
filter
- rejected: all other rows
}
\description{
Filter table rows based on a logical negation of a filter (i.e. NOT filter1)
}
