% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Perform the entire feature transformation process.}
\usage{
transform(data_sources, all_feature_json_filenames)
}
\arguments{
\item{data_sources}{A list mapping unique table identifiers to either
the file path from which they should be read from, or the data frame itself.}

\item{all_feature_json_filenames}{A vector of file paths to the feature JSON
specifications.}
}
\value{
A data frame with the feature tables joined together
}
\description{
Perform the entire feature transformation process.
}
